==> code/processing_code/1_cleaning_2020.Rmd <==
Cleaning Part 1

Load necessary libraries
```{r, echo=FALSE, message=FALSE}
library(knitr)
library(readxl)
library(lubridate)
library(stringr)
library(tidyverse)
library(here)
```

Load the files to be cleaned.
```{r}
deaths <- read_xlsx(here("data/raw_data/AIDSVu_State_Death_2020.xlsx"), skip = 3)
us_race_pop <- read_xlsx(here("data/raw_data/DECENNIALPL2020.P2-2023-02-24T124235.xlsx"), sheet = 2)
us_sex_pop <-read_csv(here("data/raw_data/raw_data.csv"), skip = 2) # not used in this analysis
kid_pop <- read_xlsx(here("data/raw_data/Child population by age group.xlsx"))
kid_race <- read_xlsx(here("data/raw_data/Child population by race and ethnicity.xlsx"))

new_dx <- read_xlsx(here("data/raw_data/AIDSVu_State_NewDX_2020.xlsx"), skip = 3)
late_dx <- read_xlsx(here("data/raw_data/AIDSVu_State_LateDiagnoses_2020.xlsx"), skip = 3)
prev  <- read_xlsx(here("data/raw_data/AIDSVu_State_Prev_2020.xlsx"), skip = 3)
rec_care  <- read_xlsx(here("data/raw_data/AIDSVu_State_ReceiptofCare_2020.xlsx"), skip = 3)
sdoh  <- read_xlsx(here("data/raw_data/AIDSVu_State_SDOH_2020.xlsx"), skip = 3)
v_sup  <- read_xlsx(here("data/raw_data/AIDSVu_State_ViralSuppression_2020.xlsx"), skip = 3)
internet <- read_xlsx(here("data/raw_data/ACSST5Y2020.S2801-2023-04-17T232857.xlsx"), skip = 2, sheet = 2)
```

Let's start by separating `deaths` by demographic characteristic . While this data provides information on demographic characteristics such as sex, age group, and race/ethnicity, it does not provide data on all of these terms together. We'll separating by sex and race/ethnicity. This analysis will be limited to Non-Hispanic Black, Non-Hispanic White, Hispanic persons, and persons of Multiple racial backgrounds.
```{r}
deaths <- deaths %>% 
  filter(!grepl("DC", `State Abbreviation`)) %>%
  filter(!grepl("PR", `State Abbreviation`))

deaths_sex <- deaths %>%
  dplyr::select(1:4, contains("male"), -contains("and"), -ends_with("Stability")) 

deaths_race <- deaths %>%
  dplyr::select(1:4, contains(c("Black", "White", "Hispanic",  "Multiple")), -ends_with("Stability"))

summary(deaths_race)
```

Some rates are -9 due to missingness. Let's convert these to NA
```{r}
deaths_race <- deaths_race %>% 
  mutate(across(ends_with("Death Rate"), ~ case_when( .x < 0 ~ NA, TRUE ~ .x)))
```


Next, let's clean the dataframe for New diagnoses. We'll select for variables of interest and subset.
```{r}
new_dx <- new_dx %>% 
  filter(!grepl("DC", `State Abbreviation`)) %>%
  filter(!grepl("PR", `State Abbreviation`)) %>% 
  dplyr::select(1:3,
                contains(c("male", "Black", "White", "Hispanic", "Multiple")),
                -contains(c("and", "Stability")))

sex_dx <- new_dx %>% 
  dplyr::select(1:7)

new_dx <- new_dx %>% 
  dplyr::select(-contains(c("male", "Year")))
```

Moving on to the census data. I'll start by re-formatting the data from wide format for States to long, and format race/ethnicity into wide.
```{r}
glimpse(us_race_pop)

race_pop <- us_race_pop %>%  
  pivot_longer(2:54 , names_to = "Location", values_to = "Population" ) %>%  
  pivot_wider(names_from = Label, values_from = Population) 

race_pop <- race_pop %>%
  filter(!grepl("Puerto Rico", Location)) %>% 
  filter(!grepl("District", Location)) 
```

I'll select only the columns of interest, and rename some of them so they are a bit easier to work with.
```{r}
race_pop <- race_pop %>% 
  dplyr::select(c(Location, starts_with(c("Total:", "Black or African American alone", "Hispanic", "White alone", "Population of two or more races:")))) %>% 
  rename(c(Hispanic = `Hispanic or Latino`,
         White = `White alone`,
         Black = `Black or African American alone`,
         Multiple = `Population of two or more races:`))
```

I need to remove all commas from the dataframe and convert the columns to numeric. 
```{r}
race_pop <- race_pop %>% 
  mutate_all(~ str_remove_all(., ",")) %>% 
  mutate(across(2:6, ~ as.numeric(.)))

total_pop <- race_pop %>%
  dplyr::select(1:2)
```

I also need to estimate the population older than 12. Since I cannot find a source explicitly population older than 12, I will estimate population younger than 13 this using data from Kids Count. 
```{r}
kid_pop <- kid_pop %>% 
  filter(TimeFrame == 2020) %>% 
  filter(`Age group` %in% c("0 to 4", "5 to 11", "12 to 14")) %>% 
  mutate(Data = as.numeric(Data))

pop_less13 <- kid_pop %>% 
  group_by(Location) %>% 
  mutate(`Age group` = ifelse(`Age group` == '12 to 14', '12', `Age group`)) %>%
  group_by(`Age group`) %>%
  mutate(`Data` = ifelse(`Age group` == '12', `Data` * 0.333, `Data`)) %>%
  group_by(Location) %>% 
  summarise(Less13 = sum(Data))
```

Then, I'll calculate the counts of males and females in each location using percentage estimates from the KFF and removing the population under 13 years old.
```{r}
new_pop <- list(total_pop, pop_less13, us_sex_pop)

new_pop <- reduce(new_pop, inner_join, by = join_by(Location))

new_pop <- new_pop %>% 
  mutate(Male_Pop = round(Male * `Total:`),
  Female_Pop = round(Female * `Total:`),
  Female_kid_pop = round(Female * Less13),
  Male_kid_pop = round(Male * Less13)) 
  
new_pop <- new_pop %>%
  filter(!grepl("District", Location)) %>%
  filter(!grepl("United", Location)) %>% 
  filter(!grepl("Puerto Rico", Location))

df_pop <- new_pop %>% 
  group_by(Location) %>% 
  summarize(pop_male = sum(Male_Pop - Male_kid_pop),
            pop_female = sum(Female_Pop - Female_kid_pop))
```

Merging dataframes. 
```{r}
sex_df <- full_join(deaths_sex, df_pop, by = join_by(State == Location))

sex_df <- full_join(sex_df, sex_dx)
```
We need to repeat the steps above, but with race and ethnicity. We'll load the file containing information on the racial/ethnic make up of the population under 18. This contains a lot of data, so we'll be sure to only select the data we need. 
```{r}
glimpse(kid_race)

kid_race <- kid_race %>% 
  filter(TimeFrame == 2020,
         DataFormat == "Percent") %>% 
  mutate(Data = as.numeric(Data)) %>% 
  filter(Race %in% c("Hispanic or Latino", "Non-Hispanic Black alone", "Non-Hispanic White alone", "Non-Hispanic Two or More Race Groups")) 

kid_race <- kid_race %>%
  pivot_wider(Location, names_from = Race, values_from = Data) %>% 
  rename(c(Hispanic_k = `Hispanic or Latino`,
         White_k = `Non-Hispanic White alone`,
         Black_k = `Non-Hispanic Black alone`,
         Multiple_k = `Non-Hispanic Two or More Race Groups`))

kid_race <- inner_join(kid_race, pop_less13)
```

Let's estimate the number of children under 13 by race. We'll use mutate and across to provide these calculations for each columns for race/ ethnicity. Then, we'll merge and subtract the population under 13 years old from the analysis.
```{r}
kid_race <- kid_race %>%
  mutate(across(2:5, ~ round( . * Less13)))

race_pop <- inner_join(race_pop, kid_race)

race_pop <- race_pop %>% 
  group_by(Location) %>% 
  summarize(Black_pop = sum(Black - Black_k),
            Hispanic_pop = sum(Hispanic - Hispanic_k),
            White_pop = sum(White - White_k),
            Multiple_pop = sum(Multiple - Multiple_k))
```

Merging datasets. 
```{r}
race_df <- list(deaths_race, late_dx, new_dx)

race_df <- reduce(race_df, inner_join)

race_df <- inner_join(race_df, race_pop, by = join_by(State == Location))
```

Cleaning the file `Received Care`. I do some cleaning and merge appropriate dataframes, and use some base R to add a couple columns to `sex_df`.
```{r}
rec_care <- rec_care %>%
  filter(!grepl("DC", `State Abbreviation`)) %>%
  filter(!grepl("PR", `State Abbreviation`)) %>% 
  dplyr::select(1:3, contains("Percent of")) %>% 
  dplyr::select(1:3, contains(c("Hispanic", "Black", "White", "Multiple","male")))

sex_df$Male_rec_care_percent <- rec_care$`Percent of Males Living with HIV who Received HIV Care`
sex_df$Female_rec_care_percent <- rec_care$`Percent of Females Living with HIV who Received HIV Care`

rec_care <- rec_care %>%
  dplyr::select(1:3, contains(c("Hispanic", "Black", "White", "Multiple")))

race_df <- inner_join(race_df, rec_care)
```
Repeat for file on `Viral Suppression`
```{r}
v_sup <- v_sup %>%
  filter(!grepl("DC", `State Abbreviation`)) %>%
  filter(!grepl("PR", `State Abbreviation`)) %>% 
  dplyr::select(1:3, contains("Percent of")) %>% 
  dplyr::select(1:3, contains(c("Hispanic", "Black", "White", "Multiple","male")))

sex_df$`Male_viral_sup_percent` <- v_sup$`Percent of Males Living with HIV who were Virally Suppressed`
sex_df$`Female_viral_sup_percent` <- v_sup$`Percent of Females Living with HIV who were Virally Suppressed`

v_sup <- v_sup %>%
  dplyr::select(1:3, contains(c("Hispanic", "Black", "White", "Multiple")))
race_df <- inner_join(race_df, v_sup)
```

Cleaning Prevalence file
```{r}
prev <- prev %>% 
  filter(!grepl("DC", `State Abbreviation`)) %>%
  filter(!grepl("PR", `State Abbreviation`)) 

prev_sex <- prev %>%
  dplyr::select(1:4, contains("male"),
                -contains("and"),
                -ends_with("Stability")) 

sex_df <- full_join(sex_df, prev_sex)

prev_race <-   prev %>%
  dplyr::select(1:4, contains(c("Black", "White", "Hispanic",  "Multiple")),
                -contains("and"),
                -ends_with("Stability")) 

race_df <- full_join(race_df, prev_race)
```


Some quick cleaning of the social determinants of health file, and merging this with our two results dataframes `sex_df` and `race_df`.
```{r}
sdoh <- sdoh %>% 
  filter(!grepl("DC", `State Abbreviation`)) %>% 
   filter(!grepl("PR", `State Abbreviation`)) %>% 
  mutate("Percent Less than High School Education" = 100 - `Percent High School Education`)

sex_df <- full_join(sex_df, sdoh)

race_df <- full_join(race_df,sdoh)
```

```{r}
no_internet <- internet %>% 
  dplyr::select(c(1, "Without an Internet subscription")) %>% 
  slice(c(-1:-8, -48:-52,-263:-267)) %>% 
  filter(!grepl(",", `Without an Internet subscription`)) %>% 
  rename(`Percent Without an Internet subscription` = `Without an Internet subscription`)
```

This dataframe has some issues. Let's fill in the state names...
```{r}
# create a vector of state names
states <- state.name

# repeat each state name five times
states <- states %>%
  rep(each = 4)

# sort the repeated state names in alphabetical order
states <- sort(states)

no_internet$State<- states 

no_internet <- no_internet %>% 
  dplyr::select(State, `Percent Without an Internet subscription`) %>% 
  drop_na()

no_internet$`Percent Without an Internet subscription` <- gsub("%", "", no_internet$`Percent Without an Internet subscription`) 

no_internet <- no_internet %>%
  mutate(`Percent Without an Internet subscription` = as.numeric(`Percent Without an Internet subscription`))
```

```{r}
race_df <- full_join(race_df, no_internet)
sex_df <- full_join(sex_df, no_internet)
```
We have some more cleaning to, but let's save these files for now. 
```{r}
saveRDS(sex_df, here("data/processed_data/sex_df.rds"))
saveRDS(race_df, here("data/processed_data/race_df.rds"))
```
