==> code/processing_code/2_cleaning_race.Rmd <==
```{r, echo=FALSE, message=FALSE}
library(knitr)
library(tidyverse)
library(here)
```

```{r}
race_df <- readRDS(here("data/processed_data/race_df.rds"))
race_df <-  race_df %>%  replace(is.na("Black Death Cases"), 0) %>% 
   replace(is.na("White Death Cases"), 0) 
```

Next, let's find the expected number of HIV deaths. First, we'll calculate lambda. 
```{r}
lambda_black <- race_df %>%
  drop_na(`Black Death Cases`) %>% 
  summarize(lambda_black = sum(`Black Death Cases`)/sum(Black_pop)) %>%
  as.numeric()

lambda_white <- race_df %>%
  drop_na(`White Death Cases`) %>% 
  summarize(lambda_white = sum(`White Death Cases`)/sum(White_pop)) %>%
  as.numeric()
```


```{r}
black_df <- race_df %>% 
  dplyr::select(1:3, contains("Black"), 50:64) %>% 
  mutate(black_death_rate = round(`Black Death Cases`/Black_pop*1e5,1))

black_df <- black_df %>% 
    group_by(State) %>% 
 rename("Death Rate (AIDSVu)" = `Black Death Rate`,
        "Population" = Black_pop,
        "Death Cases" = "Black Death Cases",
        "New Diagnoses Rate" = "New Diagnoses Black Rate",
        "New Diagnoses Cases" = "New Diagnoses Black Cases",
        "Percentage Diagnosed Late" = `Percent of Black Persons Diagnosed Late`,
        "Number of Persons Diagnosed Late" = `Number of Black Persons Diagnosed Late`,
        "Percentage Received Care" = `Percent of Black Persons Living with HIV who Received HIV Care`,
        "Percentage Virally Suppressed" = `Percent of Black Persons Living with HIV who were Virally Suppressed`,
        "HIV Prevalence Rate" = "Black Rate",
        "Prevalent HIV Cases" = "Black Cases",
        "Death Rate (estimated)" = black_death_rate) %>% 
  mutate(Race = "Black",
        "Expected Death Cases" = round(lambda_black * Population),
        "Expected Death Rate (Race)" = round(`Expected Death Cases`/`Population` *1e5,1)) %>% 
  ungroup()

white_df <- race_df %>% 
  dplyr::select(1:3, contains("White"), 50:64) %>% 
   mutate(white_death_rate = round(`White Death Cases`/White_pop*1e5,1))

white_df <- white_df %>% 
  group_by(State) %>% 
 rename("Death Rate (AIDSVu)" = `White Death Rate`,
        "Population" = White_pop,
        "Death Cases" = "White Death Cases",
        "New Diagnoses Rate" = "New Diagnoses White Rate",
        "New Diagnoses Cases" = "New Diagnoses White Cases",
        "Percentage Diagnosed Late" = `Percent of White Persons Diagnosed Late`,
        "Number of Persons Diagnosed Late" = `Number of White Persons Diagnosed Late`,
        "Percentage Received Care" = `Percent of White Persons Living with HIV who Received HIV Care`,
        "Percentage Virally Suppressed" = `Percent of White Persons Living with HIV who were Virally Suppressed`,
        "HIV Prevalence Rate" = "White Rate",
        "Prevalent HIV Cases" = "White Cases",
        "Death Rate (estimated)" = white_death_rate) %>% 
  mutate(Race = "White",
        "Expected Death Cases" = round(lambda_white * Population),
        "Expected Death Rate (Race)" = round(`Expected Death Cases`/`Population` * 1e5, 1)) %>% 
  ungroup()

race_df <- list(black_df, white_df)

race_df <- reduce(race_df, full_join)

race_df <- race_df %>% 
  mutate(`Percent HIV Prevalence` = round(`Prevalent HIV Cases`/ `Population`*100,2))
```

Now, let's calculate lambda - null, which would assume that there is no risk difference in Expected deaths by race or location
```{r}
lambda_null <- race_df %>%
  drop_na("Death Cases") %>% 
  summarize(lambda = sum(`Death Cases`)/sum(Population)) %>%
  as.numeric()

race_df <- race_df %>% 
  mutate(`Expected Death Cases (Null)` = Population * lambda_null,
         `Difference Observed - Expected` = `Death Cases` - `Expected Death Cases`,
         `Ratio Observed/ Expected (Null)` = `Death Cases`/`Expected Death Cases (Null)`,
         `Ratio Observed/ Expected (Race)` = `Death Cases`/`Expected Death Cases`,
         `Null Expected Rate` = `Expected Death Cases (Null)`/ Population*1e5,
         `Percentage Diagnosed Late` = replace(`Percentage Diagnosed Late`, which(`Percentage Diagnosed Late` < 0), 0))
```

Let's create a variable for mortality rate ratio of Black/ White
```{r}
race_df <- race_df %>%
  group_by(State) %>%
  mutate(MRR = `Death Rate (estimated)`[Race == "Black"] / `Death Rate (estimated)`[Race == "White"]) %>%
  ungroup()

#Calculate MRR with null effect of LOCATION (since lambda was calculated using sum cases and population for each race)
race_df <- race_df %>%
  group_by(State) %>% 
  mutate(MRR_null_loc = `Expected Death Rate (Race)`[Race == "Black"] / `Expected Death Rate (Race)`[Race == "White"]) %>%
  ungroup()

# MRR with null effect of LOCATION and RACE would equal 1...
```


Save the file and now we can move on to some exploratory analysis!
```{r}
saveRDS(race_df, here("data/processed_data/race_df_alt.rds"))
```
