==> code/processing_code/3_cleaning_us_nabs.Rmd <==

This is the list of neighboring states that was obtained from a user on [GitHub](https://github.com/ubikuity/List-of-neighboring-states-for-each-US-state). After some 'investigation' (troubleshooting code), this file is not exactly what I need to create the neighbor list. What is achieved by the following code took me a few seconds to do in Excel with copy and paste... but for the sake of reproducibility, I decided it's well-worth it to some spend a little bit of time do this in R. I'm essentially replicating the dataframe, but crossing over the values on the replicated rows such that both variables contain the same information. 

```{r}
library(reshape2)
library(tidyverse)
library(here)

us_nabs <- read_csv(here("data/raw_data/neighbors-states.csv"))

us_nabs_copy <- us_nabs
```

```{r}
nabs <- full_join(us_nabs, us_nabs_copy, by = join_by(StateCode == NeighborStateCode, NeighborStateCode == StateCode))
```

```{r}
saveRDS(nabs, here("data/processed_data/us_nabs.rds"))
```
